Rename Driver to VelocityLanguageDriver Fixes gh-65
