+++ src/main/java/org/mybatis/scripting/velocity/Driver.java
@@ -15,50 +15,9 @@
  */
 package org.mybatis.scripting.velocity;
 
-import org.apache.ibatis.executor.parameter.ParameterHandler;
-import org.apache.ibatis.mapping.BoundSql;
-import org.apache.ibatis.mapping.MappedStatement;
-import org.apache.ibatis.mapping.SqlSource;
-import org.apache.ibatis.parsing.XNode;
-import org.apache.ibatis.scripting.LanguageDriver;
-import org.apache.ibatis.scripting.defaults.DefaultParameterHandler;
-import org.apache.ibatis.session.Configuration;
-
-public class Driver implements LanguageDriver {
-
-  /**
-   * Default constructor.
-   */
-  public Driver() {
-    this(VelocityLanguageDriverConfig.newInstance());
-  }
-
-  /**
-   * Constructor.
-   *
-   * @param driverConfig
-   *          a language driver configuration
-   * @since 2.1.0
-   */
-  public Driver(VelocityLanguageDriverConfig driverConfig) {
-    VelocityFacade.initialize(driverConfig);
-  }
-
-  @Override
-  public ParameterHandler createParameterHandler(MappedStatement mappedStatement, Object parameterObject,
-      BoundSql boundSql) {
-    return new DefaultParameterHandler(mappedStatement, parameterObject, boundSql);
-  }
-
-  @Override
-  public SqlSource createSqlSource(Configuration configuration, XNode script, Class<?> parameterTypeClass) {
-    return new SQLScriptSource(configuration, script.getNode().getTextContent(),
-        parameterTypeClass == null ? Object.class : parameterTypeClass);
-  }
-
-  @Override
-  public SqlSource createSqlSource(Configuration configuration, String script, Class<?> parameterTypeClass) {
-    return new SQLScriptSource(configuration, script, parameterTypeClass == null ? Object.class : parameterTypeClass);
-  }
-
+/**
+ * @deprecated Since 2.1.0, recommend to use the {@link VelocityLanguageDriver}.
+ */
+@Deprecated
+public class Driver extends VelocityLanguageDriver {
 }
