+++ src/site/xdoc/index.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-       Copyright 2012-2018 the original author or authors.
+       Copyright 2012-2019 the original author or authors.
 
        Licensed under the Apache License, Version 2.0 (the "License");
        you may not use this file except in compliance with the License.
@@ -57,7 +57,7 @@
 <configuration>
   ...
   <typeAliases>
-    <typeAlias alias="velocity" type="org.mybatis.scripting.velocity.Driver"/>
+    <typeAlias alias="velocity" type="org.mybatis.scripting.velocity.VelocityLanguageDriver"/>
   </typeAliases>
   ...
 </configuration>]]></source>
